pitch_ai_mvp/requirements.txt

ultralytics==8.3.5
opencv-python
numpy
scipy
filterpy
streamlit
ffmpeg-python
